Add git-grok-middle-pr label also when creating the PRs initially, when grokking a large stack of multiple commits (#16)
## Summary

Previously, we started creating git-grok-middle-pr label for middle PRs
when updating them. It eventually delivered the label everywhere
consistently, BUT in the very beginning, when grokking multiple new
commits at once, it created the PRs without this label (and then added
it in a separate gh request), which could confuse GitHub Actions checks
related to the presence of git-grok-middle-pr label.

Also, introducing a mechanism to pass some serialized data to the stage
when a PR is created initially (previously it was hard to achieve, since
we did not have any stable bit of information about the current commit
processed in the interactive rebase). Currently, this data is like "3/5"
where 3 is the number of the current PR in the stack and 5 is the total
number of PRs in the stack (it's used to assign the label).

## PRs in the Stack
- #17
- ➡ #16

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)