Partially switch to atomic multi-branch push (#29)
## Summary

Adding an engine which is able to push multiple branches (commits)
atomically. For now, it still does most of the pushes sequentially, but
using the parallel engine (with commits list size 1).

The problem is that multi-branch push can only be used safely when no
commits reordering happened in the stack (otherwise, GitHub may mark
some of intermediate PRs as merged prematurely, and there is no way out
of it). That's why in the current PR, we do not utilize it much yet. It
will be treated in the next PRs.

## How was this tested?

CI.

## PRs in the Stack
- #31
- #30
- ➡ #29

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)