Improve closed PRs reopening corner cases (#30)
## Summary

GitHub is "too smart", so sometimes it marks the PRs as merged or as
closed on its own, right after the PR head branch is force-pushed:
1. If the head branch is force-pushed, and GitHub thinks that all the
code is merged to the upstream already, it marks the PR as merged.
2. If the head branch is force-pushed the way the PR has 0 commits in
it, GitHub closes that PR and makes it unopenable.
3. When the PR is closed anyhow, and then the head branch is
force-pushed to some new commit hash, then GitHub refuses to reopen that
PR (see https://github.com/isaacs/github/issues/361 for details).

We used to cover the case (1) (git-grok was just creating a new PR). Now
we cover the case 2 (the same way as case 1) and case 3 (with a hack
explained at https://github.com/isaacs/github/issues/361).

## How was this tested?

```
git grok
```

All those PRs were closed, then randomly reordered, and git-grok rerun:

![CleanShot 2025-04-13 at 23 05
43@2x](https://github.com/user-attachments/assets/c734bb3a-29a7-465f-b882-69b7f22a6151)

## PRs in the Stack
- #31
- ➡ #30

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)