Push multiple branches in the stack atomically if no commits reordering happened (#31)
## Summary

Git supports "atomic push" of multiple branches. We now use it in
git-grok:

It is faster than pushing one by one.
It is possible that it improves CODEOWNERS-based reviewers addition when
a large stack is reordered or rebased.

## How was this tested?

CI (there was a test that covers the usecase with reordering).

## PRs in the Stack
- ➡ #31

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)