Severely improve atomic push probability for large stacks (#32)
## Summary

Atomic multi-branch push generally works, but there was a drawback: if
you git-grok a newly added commit, then pull (even when nothing
changed!), then run git-grok again, then atomic multi-branch push didn't
trigger!

This was because when pulling, for the top commit, Git implicitly
overwrites its branch to the remote branch (which doesn't yet has the
commit message with PR url), so we were losing the PR URL for
atomic-push enablement heuristics purposes.

Now, we switch to branch-chain comparison: getting the chain of branches
in the local commits stack, getting the chain of branches in the remote
stack, and if they are equal (i.e. no reordering happened) - enabling
the atomic push.

Also, adding more unit tests to cover this exact case (when the pull is
a no-op and when a pull actually delivered someone else's commit under
your stack).

## How was this tested?

```
python3 tests/test_push_atomic.py
```

## PRs in the Stack
- ➡ #32

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)